added loader and sweet alert

+++ src/pages/MyPlant.jsx
@@ -1,12 +1,38 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Link, useLoaderData } from 'react-router';
+import Swal from 'sweetalert2';
 
 const MyPlant = () => {
     const myPlantsData = useLoaderData()
+    const [myPlants, setMyPlants] = useState(myPlantsData)
     const handleDelete = (id) => {
-        fetch(`http://localhost:3000/plant/${id}`, {
-            method: 'DELETE'
-        }).then(res => res.json()).then(data => console.log(data))
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:3000/plant/${id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.deletedCount == 1) {
+                            Swal.fire({
+                                title: "Deleted!",
+                                text: "Your file has been deleted.",
+                                icon: "success"
+                            })
+                        }
+                    })
+                const filteredPlant = myPlants.filter(plant => plant._id !== id);
+                setMyPlants(filteredPlant)
+            }
+        });
     }
     return (
         <div className="overflow-x-auto">
@@ -21,7 +47,7 @@ const MyPlant = () => {
                 </thead>
                 <tbody>
                     {
-                        myPlantsData.map(plant => <tr key={plant._id}>
+                        myPlants.map(plant => <tr key={plant._id}>
                             <td >
                                 <div className="flex items-center gap-3">
                                     <div className="avatar">
